+++ content/pages/kevin/index.mdx
@@ -5,8 +5,9 @@ slug: "/kevin"
 import ProfileHeader from "../../../src/components/profile-header"
 import Button from "../../../src/components/button"
 import AuthorProjects from "../../../src/components/author-projects"
+import KevinProfileImage from "./kevin.jpeg"
 
-<ProfileHeader name="kevin.jpeg" title="Kevin">
+<ProfileHeader image={KevinProfileImage} title="Kevin">
     <Button color="#242424" url="https://github.com/kevmca"> GitHub </Button>
     <Button color="#132CB5" url="https://www.linkedin.com/in/kevin-mcandrew-298472177/"> LinkedIn </Button>
 </ ProfileHeader>
