+++ content/pages/michael/index.mdx
@@ -5,8 +5,9 @@ slug: "/michael"
 import ProfileHeader from "../../../src/components/profile-header"
 import Button from "../../../src/components/button"
 import AuthorProjects from "../../../src/components/author-projects"
+import MichaelProfileImage from "./michael.jpeg"
 
-<ProfileHeader name="michael.jpeg" title="Michael">
+<ProfileHeader image={MichaelProfileImage} title="Michael">
     <Button color="#242424" target="_blank" url="https://github.com/mmcatcd"> GitHub </Button>
     <Button color="#F73131" target="_blank" url="https://www.mcandrew.dev/"> Personal site </Button>
     <Button color="#132CB5" target="_blank" url="https://www.linkedin.com/in/michael-mcandrew-280a0476/"> LinkedIn </Button>
