--- src/components/project-picture.tsx
@@ -1,14 +1,0 @@
-/** @jsx jsx */
-import { jsx, Link } from "theme-ui"
-
-const ProjectPicture = (props) => {
-
-  return (
-    <img 
-        src={props.src}
-        sx={{ width: 150, objectFit: "cover", height: 150, borderRadius: 150/ 2, border: "1px solid grey"}}
-    />
-  )
-}
-
-export default ProjectPicture
